\name{Discrete Laplace package}
\alias{disclap-package}
\alias{disclap}
\docType{package}
\title{Discrete Laplace Family}
\description{Discrete Laplace Family for models such as a generalized linear model.}
\details{
DiscreteLaplace()
ddisclap(x, p)
pdisclap(x, p, lower.tail = TRUE)
rdisclap(n, p)
}
\author{
Mikkel Meyer Andersen and Poul Svante Eriksen

Maintainer: Mikkel Meyer Andersen <mikl@math.aau.dk>
}
\keyword{package}
\seealso{
\code{\link{DiscreteLaplace}}
\code{\link{ddisclap}}
\code{\link{pdisclap}}
\code{\link{rdisclap}}
}


